migrate/requirements_py3.txt
altgraph==0.10.2
backports-abc==0.5
bdist-mpkg==0.5.0
# bonjour-py==0.0.3
boto==2.48.0
boto3==1.9.227
botocore==1.12.227
certifi==2020.6.20
configobj==5.0.6
docutils==0.15.2
future==0.18.2
futures==3.1.1
jmespath==0.9.4
macholib==1.5.1
# matplotlib==1.3.1
modulegraph==0.10.4
nose==1.3.7
numpy==1.19.1
py2app==0.7.3
# pyobjc-core==2.5.1
# pyobjc-framework-Accounts==2.5.1
# pyobjc-framework-AddressBook==2.5.1
# pyobjc-framework-AppleScriptKit==2.5.1
# pyobjc-framework-AppleScriptObjC==2.5.1
# pyobjc-framework-Automator==2.5.1
# pyobjc-framework-CFNetwork==2.5.1
# pyobjc-framework-Cocoa==2.5.1
# pyobjc-framework-Collaboration==2.5.1
# pyobjc-framework-CoreData==2.5.1
# pyobjc-framework-CoreLocation==2.5.1
# pyobjc-framework-CoreText==2.5.1
# pyobjc-framework-DictionaryServices==2.5.1
# pyobjc-framework-EventKit==2.5.1
# pyobjc-framework-ExceptionHandling==2.5.1
# pyobjc-framework-FSEvents==2.5.1
# pyobjc-framework-InputMethodKit==2.5.1
# pyobjc-framework-InstallerPlugins==2.5.1
# pyobjc-framework-InstantMessage==2.5.1
# pyobjc-framework-LatentSemanticMapping==2.5.1
# pyobjc-framework-LaunchServices==2.5.1
# pyobjc-framework-Message==2.5.1
# pyobjc-framework-OpenDirectory==2.5.1
# pyobjc-framework-PreferencePanes==2.5.1
# pyobjc-framework-PubSub==2.5.1
# pyobjc-framework-QTKit==2.5.1
# pyobjc-framework-Quartz==2.5.1
# pyobjc-framework-ScreenSaver==2.5.1
# pyobjc-framework-ScriptingBridge==2.5.1
# pyobjc-framework-SearchKit==2.5.1
# pyobjc-framework-ServiceManagement==2.5.1
# pyobjc-framework-Social==2.5.1
# pyobjc-framework-SyncServices==2.5.1
# pyobjc-framework-SystemConfiguration==2.5.1
# pyobjc-framework-WebKit==2.5.1
# pyOpenSSL==0.13.1
pyparsing==2.0.2
python-dateutil==1.5
pytz==2013.7
s3transfer==0.2.1
scipy==1.5.2
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
# xattr==0.6.4
openpyxl
foundry
transaction
akhet
bottle
html5lib
pyyaml